+++ code/src/reducers/todo.js
@@ -1,0 +1,27 @@
+/* eslint-disable max-len */
+import { createSlice } from '@reduxjs/toolkit';
+
+const todo = createSlice({
+  name: 'todo',
+  initialState: {
+    items: [
+      { id: 1, text: 'Watch video on actions & reducers', complete: true },
+      { id: 2, text: 'Follow redux codealong', complete: true },
+      { id: 3, text: 'Fork weekly assignment', complete: true },
+      { id: 4, text: 'Create a todo app', complete: false }
+    ]
+  },
+  reducers: {
+    toggleItem: (store, action) => {
+      console.log(store);
+      console.log(action);
+      store.items.forEach((item) => {
+        if (item.id === action.payload) {
+          item.isComplete = !item.isComplete
+        }
+      });
+    }
+  }
+});
+
+export default todo;
